+++ gestor/src/main/java/com/floresta/gestor/service/VentasService.java
@@ -33,7 +33,7 @@ public class VentasService {
 		if (id == null) return false;
 
 	    return ventaRepository.findById(id).map(v -> {
-	      Integer pedidoId = v.getIdPedido();// o getIdPedido()
+	      Long pedidoId = v.getIdPedido();// o getIdPedido()
 	      pedidoRepository.deleteById(pedidoId);    // ← borra PADRE → cascada borra Venta+items
 	      return true;
 	    }).orElse(false);
